review api stage 2

+++ src/main/java/com/example/wbdvsu19projectserver/repositories/EditorRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.repository.CrudRepository;
  * @author Xuan Kong
  * @Date 2019-07-27.
  */
-public interface EditorRepository extends CrudRepository<Editor,Long> {
+public interface EditorRepository extends CrudRepository<Editor,Integer> {
 }
