+++ src/main/java/com/example/wbdvsu19projectserver/repositories/ReviewRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.repository.CrudRepository;
  * @author Xuan Kong
  * @Date 2019-07-27.
  */
-public interface ReviewRepository extends CrudRepository<Review,Long> {
+public interface ReviewRepository extends CrudRepository<Review,Integer> {
 }
